sw(cgh)	R Documentation

Perform the Smith-Waterman Algorithm

Description

Perform the Smith-Waterman algorithm on a vector of real values.

Usage


  sw(x, max.nIslands = NULL, trace = FALSE)
  

Arguments

x 	 a vector of real values

max.nIslands 	 the number of iterations of the algorithm performed. Each iteration finds the next highest-scoring 'island' of positive values. Set to NULL to find all islands

trace 	 print verbose output if TRUE

Details

The Smith-Waterman algorithm detects 'islands' of positive scores in a vector of real values. The input values should have a negative mean. The algorithm can be used to identify regions of copy number change in microarray fluorescence logratios, once the logratios have been adjusted for sign and a suitable threshold value subtracted to ensure a negative mean: see sw.threshold

Value

x 	 the input vector

s 	 a numeric vector containing the partial sums after one iteration of the Smith-Waterman algorithm

score 	 a numeric vector of island scores

start 	 a numeric vector of indices identifying the start of each island

length 	 a numeric vector of island lengths

Author(s)

T.S.Price

References

Smith TF, Waterman MS. Identification of common molecular subsequences. J Mol Biol. 1981;147(1):195-7.

See Also

sw.threshold sw.perm.test sw.rob sw.plot

Examples


## simluate vector of logratios
set.seed(3)
logratio <- c(rnorm(20) - 1, rnorm(20))

## invert sign of values and subtract threshold to ensure negative mean
x <- sw.threshold(logratio, function(x) median(x) + .2 * mad(x), sign = -1)

## perform Smith-Waterman algorithm
sw(x, trace = TRUE)
  

[Package Contents]